Make game ignore last button in command sequence
This feature can be disabled with a new toggle setting.

The `before_GameState_Update` hook, renamed to `before_GameState_Loop`,
is now before all the functions in the main loop of Graph_ThreadEntry:
this allows to update rInputCtx and trigger commands before the game
processes inputs.

The command's last input is stored after the command executes, and
is then removed from the game's input-related structs before their
effect is applied, via the `GameButtonInputs` hook.

The `SaveMenuIgnoreOpen` patch has been removed, as it would now be
redundant.